refactor: extract `loadConfig` into separate file
