+++ lib/index.js
@@ -1,7 +1,7 @@
-import { cosmiconfig } from 'cosmiconfig'
 import debug from 'debug'
 import inspect from 'object-inspect'
 
+import { loadConfig } from './loadConfig.js'
 import { PREVENTED_EMPTY_COMMIT, GIT_ERROR, RESTORE_STASH_EXAMPLE } from './messages.js'
 import { printTaskOutput } from './printTaskOutput.js'
 import { runAll } from './runAll.js'
@@ -15,37 +15,6 @@ import { validateConfig } from './validateConfig.js'
 import { validateOptions } from './validateOptions.js'
 
 const debugLog = debug('lint-staged')
-
-const resolveConfig = (configPath) => {
-  try {
-    return require.resolve(configPath)
-  } catch {
-    return configPath
-  }
-}
-
-const loadConfig = (configPath) => {
-  const explorer = cosmiconfig('lint-staged', {
-    searchPlaces: [
-      'package.json',
-      '.lintstagedrc',
-      '.lintstagedrc.json',
-      '.lintstagedrc.yaml',
-      '.lintstagedrc.yml',
-      '.lintstagedrc.mjs',
-      '.lintstagedrc.js',
-      '.lintstagedrc.cjs',
-      'lint-staged.config.mjs',
-      'lint-staged.config.js',
-      'lint-staged.config.cjs',
-    ],
-    loaders: {
-      '.mjs': (path) => import(path).then((module) => module.default),
-    },
-  })
-
-  return configPath ? explorer.load(resolveConfig(configPath)) : explorer.search()
-}
 
 /**
  * @typedef {(...any) => void} LogFunction
